requirements.txt
google-cloud-bigquery==3.38.0
google-cloud-bigquery-storage==2.34.0
pandas==2.3.3
db-dtypes==1.4.3
fastapi
uvicorn
google-auth
python-dotenv